HitManager changes
With how the project is being set-up now, the player will be contained in their own scene. This means that I can't have manually created references to and from the player, to outside things like enemies.
So, the player can't be the location of the HitManager anymore (as the HitManager needs to cache enemies to work (although, it could just do all that during runtime as well)).
And, the BasicBulletManager can no longer have a manual reference to the HitManager either. It still needs the HitManager of course, but now it gets and caches it during runtime, by getting a reference to it from the enemy they hit.

Also though, I'm realising now that I might want multiple scenes of enemies that I can load in, during run-time, which'd mean multiple HitManagers or atleast somekind of runtime system for adding them to a single HitManager that's already active. This is pretty silly though, cause my game is focused on the boss, there'll probably be barely any other enemies in the game. Yet, I'm considering making this big system for handling all those enemies.
I don't mind spending time on systems like that, it's satisfying; my concern is with time-management, even if I want to do it, and have the energy for it, can I justify it?
Well, I've got a lot of time right now, but it is a concern. The HitManager system probably will receive a rework eventually.