Bullet System Changes
This is a minor rework of the system. It's still the same mechanically though. System changes are related to changes of it's inheritance. And there are some new additions for the audio too.

It's a bit stinky to have it all in one big commit like this, but I do it like this because all these essentially depend on each-other. If I was to separate it into parts (like having the audio stuff as a separate commit) then there'd be compiler errors if you were to for some-reason launch off of that commit. Maybe it's still the right thing to do though, I'm not sure.