Emitter Groups
I made these to keep things tidy in the animation window (which is how I'm creating the attack patterns).
Basically, it just groups a bunch of emitters under one monobehaviour, and then you can call emit on that group to emit all groups at once.
Then there are these bonus scripts, the AnimEventLink ones. They just exist because animation events can't call on scripts on child gameobjects, and I want the groups to be on child gameobjects for organisation reasons. So, the link things just store a reference to a group and can then be called by the animation event; A middle man.