Resit Docs added
Mostly just a bunch of random notes. I reference them in the PDP though so I thought I should add them here, given that the links in that PDP will probably break one day.
Not everything from those links are accessible here though. Mainly the style vault, it would just be a waste of storage space to include all that.
It was just a bunch of image files and videos I had gathered from the internet. The mood-board and the links basically suffice as the same thing.